+++ src/daemon/get_blocks_message_skyencoder.go
@@ -4,8 +4,8 @@ package daemon
 import "github.com/skycoin/skycoin/src/cipher/encoder"
 
 // EncodeSizeGetBlocksMessage computes the size of an encoded object of type GetBlocksMessage
-func EncodeSizeGetBlocksMessage(obj *GetBlocksMessage) int {
-	i0 := 0
+func EncodeSizeGetBlocksMessage(obj *GetBlocksMessage) uint64 {
+	i0 := uint64(0)
 
 	// obj.LastBlock
 	i0 += 8
