+++ src/daemon/give_txns_message_skyencoder.go
@@ -11,13 +11,13 @@ import (
 )
 
 // EncodeSizeGiveTxnsMessage computes the size of an encoded object of type GiveTxnsMessage
-func EncodeSizeGiveTxnsMessage(obj *GiveTxnsMessage) int {
-	i0 := 0
+func EncodeSizeGiveTxnsMessage(obj *GiveTxnsMessage) uint64 {
+	i0 := uint64(0)
 
 	// obj.Transactions
 	i0 += 4
 	for _, x := range obj.Transactions {
-		i1 := 0
+		i1 := uint64(0)
 
 		// x.Length
 		i1 += 4
@@ -31,29 +31,29 @@ func EncodeSizeGiveTxnsMessage(obj *GiveTxnsMessage) int {
 		// x.Sigs
 		i1 += 4
 		{
-			i2 := 0
+			i2 := uint64(0)
 
 			// x
 			i2 += 65
 
-			i1 += len(x.Sigs) * i2
+			i1 += uint64(len(x.Sigs)) * i2
 		}
 
 		// x.In
 		i1 += 4
 		{
-			i2 := 0
+			i2 := uint64(0)
 
 			// x
 			i2 += 32
 
-			i1 += len(x.In) * i2
+			i1 += uint64(len(x.In)) * i2
 		}
 
 		// x.Out
 		i1 += 4
 		{
-			i2 := 0
+			i2 := uint64(0)
 
 			// x.Address.Version
 			i2++
@@ -67,7 +67,7 @@ func EncodeSizeGiveTxnsMessage(obj *GiveTxnsMessage) int {
 			// x.Hours
 			i2 += 8
 
-			i1 += len(x.Out) * i2
+			i1 += uint64(len(x.Out)) * i2
 		}
 
 		i0 += i1
@@ -89,7 +89,7 @@ func EncodeGiveTxnsMessage(buf []byte, obj *GiveTxnsMessage) error {
 	}
 
 	// obj.Transactions length check
-	if len(obj.Transactions) > math.MaxUint32 {
+	if uint64(len(obj.Transactions)) > math.MaxUint32 {
 		return errors.New("obj.Transactions length exceeds math.MaxUint32")
 	}
 
@@ -114,7 +114,7 @@ func EncodeGiveTxnsMessage(buf []byte, obj *GiveTxnsMessage) error {
 		}
 
 		// x.Sigs length check
-		if len(x.Sigs) > math.MaxUint32 {
+		if uint64(len(x.Sigs)) > math.MaxUint32 {
 			return errors.New("x.Sigs length exceeds math.MaxUint32")
 		}
 
@@ -135,7 +135,7 @@ func EncodeGiveTxnsMessage(buf []byte, obj *GiveTxnsMessage) error {
 		}
 
 		// x.In length check
-		if len(x.In) > math.MaxUint32 {
+		if uint64(len(x.In)) > math.MaxUint32 {
 			return errors.New("x.In length exceeds math.MaxUint32")
 		}
 
@@ -156,7 +156,7 @@ func EncodeGiveTxnsMessage(buf []byte, obj *GiveTxnsMessage) error {
 		}
 
 		// x.Out length check
-		if len(x.Out) > math.MaxUint32 {
+		if uint64(len(x.Out)) > math.MaxUint32 {
 			return errors.New("x.Out length exceeds math.MaxUint32")
 		}
