+++ src/visor/blockdb/block_skyencoder.go
@@ -11,8 +11,8 @@ import (
 )
 
 // EncodeSizeBlock computes the size of an encoded object of type Block
-func EncodeSizeBlock(obj *coin.Block) int {
-	i0 := 0
+func EncodeSizeBlock(obj *coin.Block) uint64 {
+	i0 := uint64(0)
 
 	// obj.Head.Version
 	i0 += 4
@@ -38,7 +38,7 @@ func EncodeSizeBlock(obj *coin.Block) int {
 	// obj.Body.Transactions
 	i0 += 4
 	for _, x := range obj.Body.Transactions {
-		i1 := 0
+		i1 := uint64(0)
 
 		// x.Length
 		i1 += 4
@@ -52,29 +52,29 @@ func EncodeSizeBlock(obj *coin.Block) int {
 		// x.Sigs
 		i1 += 4
 		{
-			i2 := 0
+			i2 := uint64(0)
 
 			// x
 			i2 += 65
 
-			i1 += len(x.Sigs) * i2
+			i1 += uint64(len(x.Sigs)) * i2
 		}
 
 		// x.In
 		i1 += 4
 		{
-			i2 := 0
+			i2 := uint64(0)
 
 			// x
 			i2 += 32
 
-			i1 += len(x.In) * i2
+			i1 += uint64(len(x.In)) * i2
 		}
 
 		// x.Out
 		i1 += 4
 		{
-			i2 := 0
+			i2 := uint64(0)
 
 			// x.Address.Version
 			i2++
@@ -88,7 +88,7 @@ func EncodeSizeBlock(obj *coin.Block) int {
 			// x.Hours
 			i2 += 8
 
-			i1 += len(x.Out) * i2
+			i1 += uint64(len(x.Out)) * i2
 		}
 
 		i0 += i1
@@ -131,7 +131,7 @@ func EncodeBlock(buf []byte, obj *coin.Block) error {
 	}
 
 	// obj.Body.Transactions length check
-	if len(obj.Body.Transactions) > math.MaxUint32 {
+	if uint64(len(obj.Body.Transactions)) > math.MaxUint32 {
 		return errors.New("obj.Body.Transactions length exceeds math.MaxUint32")
 	}
 
@@ -156,7 +156,7 @@ func EncodeBlock(buf []byte, obj *coin.Block) error {
 		}
 
 		// x.Sigs length check
-		if len(x.Sigs) > math.MaxUint32 {
+		if uint64(len(x.Sigs)) > math.MaxUint32 {
 			return errors.New("x.Sigs length exceeds math.MaxUint32")
 		}
 
@@ -177,7 +177,7 @@ func EncodeBlock(buf []byte, obj *coin.Block) error {
 		}
 
 		// x.In length check
-		if len(x.In) > math.MaxUint32 {
+		if uint64(len(x.In)) > math.MaxUint32 {
 			return errors.New("x.In length exceeds math.MaxUint32")
 		}
 
@@ -198,7 +198,7 @@ func EncodeBlock(buf []byte, obj *coin.Block) error {
 		}
 
 		// x.Out length check
-		if len(x.Out) > math.MaxUint32 {
+		if uint64(len(x.Out)) > math.MaxUint32 {
 			return errors.New("x.Out length exceeds math.MaxUint32")
 		}
